速度測定

+++ middleware/extract_info.py
@@ -1,28 +1,55 @@
 import pyworld as pw
 from pydub import AudioSegment
 from pydub.silence import split_on_silence
+from google.cloud import speech_v1p1beta1
+from google.cloud.speech_v1p1beta1 import enums
+from mutagen.mp3 import MP3
 import numpy as np
 
 import os
+import io
 
-def main(filename, cfg):
-    print('analyze')
+def extract_speed(file_path):
+    audio = MP3(file_path)
+    speaking_time = audio.info.length
 
-    file_path = cfg['FILE_PATH'] + '/' + filename
+    client = speech_v1p1beta1.SpeechClient()
 
-    res = {
-        'db': [],
-        'db_max': [],
-        'f0': [],
-        # 'sp': [],
-        # 'ap': []
+    language_code = "ja-JP"
+    sample_rate_hertz = 44100
+    encoding = enums.RecognitionConfig.AudioEncoding.MP3
+    config = {
+        "language_code": language_code,
+        "sample_rate_hertz": sample_rate_hertz,
+        "encoding": encoding,
     }
 
-    if not os.path.exists(file_path):
-        return res
+    with io.open(file_path, "rb") as f:
+        content = f.read()
+    audio = {"content": content}
+    response = client.recognize(config, audio)
 
-    sound = AudioSegment.from_file(file_path, 'mp3')
+    transcripts = []
+    for result in response.results:
+        alternative = result.alternatives[0]
+        transcripts.append(alternative.transcript)
 
+    word_counts = [0] * len(transcripts)
+    for i, sentence in enumerate(transcripts):
+        word_counts[i] = len(sentence)
+
+    speaking_rate = []
+    for i in range(len(transcripts)):
+        speaking_rate.append(word_counts[i]/speaking_time)
+
+    res = {}
+    res['speaking_rate'] = speaking_rate
+
+    print(speaking_time)
+
+    return res
+
+def extract_pitch_and_db(sound):
     fs = sound.frame_rate
     data = np.array(sound.get_array_of_samples()).astype(np.float)
 
@@ -33,11 +60,31 @@ def main(filename, cfg):
     # sp = pw.cheaptrick(data, f0, _time, fs)  # スペクトル包絡の抽出
     # ap = pw.d4c(data, f0, _time, fs)  # 非周期性指標の抽出
 
+    res = {}
     res['db'] = db
     res['db_max'] = db_max
     res['f0'] = f0.tolist()
     # res['sp'] = sp.tolist()
     # res['ap'] = ap.tolist()
+
+    print(_time)
+    return res
+
+def main(filename, cfg):
+    print('analyze')
+
+    file_path = cfg['FILE_PATH'] + '/' + filename
+
+    res = {}
+
+    if not os.path.exists(file_path):
+        return res
+
+    sound = AudioSegment.from_file(file_path, 'mp3')
+    res_1 = extract_pitch_and_db(sound)
+    res.update(res_1)
+    res_2 = extract_speed(file_path)
+    res.update(res_2)
 
     os.remove(file_path)
